image2text/requirements.txt
pillow==11.3.0
torch==2.7.1
torchaudio==2.7.1
torchvision==0.22.1
transformers==4.53.0
fastapi==0.115.14
pydantic==2.11.7
pydantic_core==2.33.2
uvicorn==0.35.0
sentencepiece==0.2.0
tqdm==4.67.1
kaggle==1.7.4.5
evaluate==0.4.4
peft==0.16.0
bitsandbytes==0.46.1
pyyaml==6.0.2